requirements.txt
sqlalchemy
mysqlclient
marshmallow-sqlalchemy
mysql-connector-python
